+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [v0.2.2] - 2019-11-27
 ### Added
 - Allow `s.str`'s `regex` kwarg to be a pre-existing pattern returned from
   `re.compile` (#36)
@@ -49,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial commit
 
+[v0.2.2]: https://github.com/coverahealth/dataspec/compare/v0.2.1..v0.2.2
 [v0.2.1]: https://github.com/coverahealth/dataspec/compare/v0.2.0..v0.2.1
 [v0.2.0]: https://github.com/coverahealth/dataspec/compare/v0.1.post1..v0.2.0
 [v0.1.post1]: https://github.com/coverahealth/dataspec/releases/tag/v0.1.post1